refactor : action, state hook으로 분리
